Refactoring AboutPage structure
