+++ app/components/AboutPage/AdvisoryNetwork/AdvisoryNetworkImage.js
