+++ app/components/AboutPage/Team/TeamImage.js
