+++ app/components/AboutPage/Team/TeamListItem.js
